Fix small typos in lab naming for Git Workshop
This change fixes small typos in the naming of the Git Workshop.
Also rename Branching and Merging to a more suitable name,
Advanced, as branching and merging is included in Introduction.

This also adds .DS_Store files to the .gitignore file as they
should not be committed if they are generated for any reason.