Add commit lab as page one of the introduction workshop
Currently page one of the introduction workshop is an overview of
branching and merging. However, page two gets into branching
pretty quickly.

There is a benefit in having a more thorough understanding of
making changes, staging them and then committing them. Especially
as some members of the workshop will be working on this remotely
and asynchronously.

This change adds a commits lab to page one instead of the branching
and merging overview. This covers:

* Creating a first commit
* Understanding what's happening when staging changes
* Understanding the lifecycle of a file in git