Update branching and merging
Currently the branching lab contains some information that is now
in the commit lab. It is also not very structured which may cause
some difficulty to anyone who is doing the workshop remotely or
asynchronously.

This makes some structural changes to the branching lab and also
adds in a number of diagrams to help the understanding of what the
branches are doing.

The merging lab was good so this commit just updates the structure
of the lab and make it reference back to the branching lab.